pyproject.toml
[project]
name = "color-floorplan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.11"
dependencies = [
    "gltflib>=1.0.13",
    "google-auth>=2.45.0",
    "google-genai>=1.56.0",
    "mapbox-earcut>=2.0.0",
    "numpy>=2.4.0",
    "pillow>=12.0.0",
    "shapely>=2.1.2",
    "svgpathtools>=1.7.2",
]
